# Docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MyLab API (Go)
  version: 0.1.0
  description: |
    Contract-first API definition. Keep this file in sync with handlers and JSON examples.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        '200':
          description: OK

  /readyz:
    get:
      summary: Readiness check
      responses:
        "200":
          description: Ready
        "503":
          description: Not ready

  /metrics:
    get:
      summary: Prometheus metrics
      responses:
        "200":
          description: Metrics

  /v1/auth/login:
    post:
      summary: User authentication (login)
      description: |
        Authenticate user with email and password.
        Returns user information on success.
        No authentication header required for this endpoint.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User email address (case-insensitive)
                  example: admin@example.com
                password:
                  type: string
                  format: password
                  description: User password
                  example: password123
            examples:
              valid:
                summary: Valid credentials
                value:
                  email: admin@example.com
                  password: password123
              caseInsensitiveEmail:
                summary: Case-insensitive email
                value:
                  email: ADMIN@EXAMPLE.COM
                  password: password123
              missingFields:
                summary: Missing required fields
                value:
                  email: ""
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                required:
                  - ok
                  - message
                  - user_id
                  - company_id
                properties:
                  ok:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Login successful.
                  user_id:
                    type: integer
                    format: int64
                    description: User ID
                    example: 1
                  company_id:
                    type: integer
                    format: int64
                    description: Company/tenant ID
                    example: 1
                  role:
                    type: string
                    description: User role (optional, omitted if null in database)
                    example: admin
              examples:
                withRole:
                  summary: Login success with role
                  value:
                    ok: true
                    message: Login successful.
                    user_id: 1
                    company_id: 1
                    role: admin
                withoutRole:
                  summary: Login success without role
                  value:
                    ok: true
                    message: Login successful.
                    user_id: 2
                    company_id: 1
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnauthorizedError'
              examples:
                invalidCredentials:
                  summary: Wrong password or user not found
                  value:
                    ok: false
                    message: Unauthorized.
                    errors:
                      credentials: invalid
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceValidationError'
              examples:
                missingFields:
                  summary: Missing required fields
                  value:
                    ok: false
                    message: Validation failed.
                    errors:
                      email: required
                      password: required
                invalidJson:
                  summary: Invalid JSON body
                  value:
                    ok: false
                    message: Validation failed.
                    errors:
                      body: invalid JSON
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceError'

  /v1/query:
    post:
      summary: Execute restricted query (Laravel-style DSL)
      description: |
        Executes a tenant-enforced, whitelist-validated query built from a restricted
        Laravel-style query builder chain.

        The server does NOT execute raw SQL strings.
      tags:
        - Query
      security:
        - UserIdHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LaravelQueryRequest'
            examples:
              menuExample:
                summary: Query menu with LIKE and LIMIT
                value:
                  laravel_query: "table('menu as m')->select('m.id','m.menu_name')->where('m.menu_name','like','admin')->orderby('m.id','desc')->take(10)"
      responses:
        '200':
          description: OK

  /v1/crud/{table}:
    post:
      summary: Generic CRUD - create
      description: |
        Create a record in the given table. Tenant-enforced via company_id.
        Table access is controlled via env (CRUD_ALLOWED_TABLES / CRUD_DENIED_TABLES).
      tags:
        - Generic CRUD
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericCRUDCreateResponse'

  /v1/crud/{table}/{pk}:
    get:
      summary: Generic CRUD - get
      tags:
        - Generic CRUD
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
        - in: path
          name: pk
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericCRUDGetResponse'
    put:
      summary: Generic CRUD - update
      tags:
        - Generic CRUD
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
        - in: path
          name: pk
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericCRUDWriteResponse'
    patch:
      summary: Generic CRUD - patch
      tags:
        - Generic CRUD
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
        - in: path
          name: pk
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericCRUDWriteResponse'
    delete:
      summary: Generic CRUD - delete
      tags:
        - Generic CRUD
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
        - in: path
          name: pk
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericCRUDWriteResponse'

  /v1/crud/{table}/select:
    post:
      summary: Generic CRUD - select
      description: Safe select with where/like/order_by/page/per_page.
      tags:
        - Generic CRUD
      parameters:
        - in: path
          name: table
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericCRUDSelectRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericCRUDSelectResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnauthorizedError'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceValidationError'
        '500':
          description: Server error
components:
  securitySchemes:
    UserIdHeader:
      type: apiKey
      in: header
      name: X-User-Id
      description: |
        Required for all /v1 endpoints.
        The API looks up the user in `users` and derives tenant context from `users.company_id`.
  schemas:
    ServiceOkResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
    ServiceValidationError:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string

    ServiceUnauthorizedError:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string

    ServiceNotFoundError:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string

    LaravelQueryRequest:
      type: object
      required:
        - laravel_query
      properties:
        laravel_query:
          type: string
          description: |
            Restricted Laravel-style query builder chain.
            The server parses and validates this DSL; raw SQL is not executed.
          example: "table('menu as m')->select('m.id','m.menu_name')->where('m.menu_name','like','admin')->orderby('m.id','desc')->take(10)"

    GenericQueryResponse:
      type: object
      required:
        - ok
        - message
        - data
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
          example: OK
        data:
          type: array
          items:
            type: object
            additionalProperties: true

    GenericCRUDCreateResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        table:
          type: string
        pk:
          description: Primary key value
          oneOf:
            - type: string
            - type: integer

    GenericCRUDWriteResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        table:
          type: string
        pk:
          type: string

    GenericCRUDGetResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        data:
          type: object
          additionalProperties: true

    GenericCRUDOrderBy:
      type: object
      properties:
        field:
          type: string
        dir:
          type: string
          enum: [asc, desc]

    GenericCRUDSelectRequest:
      type: object
      properties:
        select:
          type: array
          items:
            type: string
        where:
          type: object
          additionalProperties: true
        like:
          type: object
          additionalProperties: true
        order_by:
          type: array
          items:
            $ref: '#/components/schemas/GenericCRUDOrderBy'
        page:
          type: integer
        per_page:
          type: integer

    GenericCRUDSelectPaging:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        has_more:
          type: boolean

    GenericCRUDSelectResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            type: object
            additionalProperties: true
        paging:
          $ref: '#/components/schemas/GenericCRUDSelectPaging'